Refactor to objects
- [ADDED] Send offer content to mail
- [CHANGED] Subscriber -> SubscriberEntity
