+++ services/emailer/EmailerQueueService.php
@@ -33,7 +33,7 @@ class EmailerQueueService
             }
 
             foreach ($ar->findAll($city['id']) as $sub) {
-                $this->queue->push(new MailJob($this->emailer, $sub->email, $sub->id));
+                $this->queue->push(new MailJob($this->emailer, $sub, $offer));
                 $this->logToDb($sub->id, $offer->title, $offer->content);
             }
         }
