+++ services/emailer/jobs/MailJob.php
@@ -3,6 +3,8 @@
 namespace app\services\emailer\jobs;
 
 use app\services\emailer\Emailer;
+use app\services\emailer\entities\OfferEntity;
+use app\services\emailer\entities\SubscriberEntity;
 use yii\queue\JobInterface;
 use yii\symfonymailer\Message;
 
@@ -12,15 +14,17 @@ class MailJob implements JobInterface
     public const STATE_INPROGRESS = 1;
     public const STATE_DONE = 2;
 
-    public function __construct(private Emailer $emailer, private string $email, private string $id)
+    public function __construct(private Emailer $emailer, private SubscriberEntity $sub, private OfferEntity $offer)
     {
     }
 
     public function execute($queue): void
     {
         $message = new Message();
-        $this->emailer->send($message, $this->email, $this->id);
-        $this->changeLogState($this->id);
+        $message->setSubject($this->offer->title);
+        $message->setTextBody($this->offer->content);
+        $this->emailer->send($message, $this->sub->email, $this->sub->id);
+        $this->changeLogState($this->sub->id);
     }
 
     public function changeLogState(int $id): void
